torchserve_TR/requirements_torchserve.txt
torch
torchvision
opencv-python
pytesseract
transformers
pillow
numpy